- Float hours, EasyScan popup, info panel, forecast fix on edit,
  watt validation, list refresh on resume
- Resolved Device constructor conflict (float hours + usage/cost fields)
- Fixed Group constructor call with new usage parameter

+++ app/src/main/java/com/example/savesomebills/DeviceStorage.java
@@ -63,8 +63,8 @@ public class DeviceStorage {
                 Device d = new Device(
                     obj.getString("name"),
                     obj.getString("groupId"),
-                    obj.getInt("onHours"),
-                    obj.getInt("standbyHours"),
+                    (float) obj.getDouble("onHours"),
+                    (float) obj.getDouble("standbyHours"),
                     obj.optInt("wattOn", 0),
                     obj.optInt("wattStandby", 0),
                     obj.optString("icon", "⚡")
